+++ src/main/java/org/gridsuite/securityanalysis/server/SecurityAnalysisController.java
@@ -16,6 +16,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import org.gridsuite.securityanalysis.server.service.SecurityAnalysisRunContext;
+import org.gridsuite.securityanalysis.server.service.SecurityAnalysisService;
+import org.gridsuite.securityanalysis.server.service.SecurityAnalysisWorkerService;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
