+++ src/main/java/org/gridsuite/securityanalysis/server/service/ActionsService.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.securityanalysis.server;
+package org.gridsuite.securityanalysis.server.service;
 
 import com.powsybl.contingency.Contingency;
 import org.springframework.beans.factory.annotation.Autowired;
