+++ src/test/java/org/gridsuite/securityanalysis/server/SecurityAnalysisControllerTest.java
@@ -10,6 +10,10 @@ import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.test.EurostagTutorialExample1Factory;
 import com.powsybl.network.store.client.NetworkStoreService;
 import com.powsybl.network.store.client.PreloadingStrategy;
+import org.gridsuite.securityanalysis.server.service.ActionsService;
+import org.gridsuite.securityanalysis.server.service.SecurityAnalysisConfigService;
+import org.gridsuite.securityanalysis.server.service.SecurityAnalysisService;
+import org.gridsuite.securityanalysis.server.service.SecurityAnalysisWorkerService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
